fix: 비밀번호 확인 함수 추가

+++ src/hooks/usePhototalkAction.tsx
@@ -2,6 +2,7 @@ import {
   useDeletePhototalkByAdmin,
   useDeletePhototalkByGuest,
 } from '@/hooks/usePhototalk';
+// import { useDeletePhototalkS3Image } from '@/hooks/useS3Image';
 import { PhotoTalk } from '@/types/phototalkType';
 import { ACTION_MODE, ActionMode, USER_MODE, UserMode } from '@/types/users';
 import { useState } from 'react';
@@ -10,12 +11,14 @@ interface usePhototalkActionProps {
   mode: UserMode;
   onEdit?: (photoTalk: PhotoTalk) => void;
   onDelete?: (photoTalk: PhotoTalk) => void;
+  refetch?: () => void;
 }
 
 export const usePhototalkAction = ({
   mode,
   onEdit,
   onDelete,
+  refetch,
 }: usePhototalkActionProps) => {
   const [isModalOpen, setModalOpen] = useState(false);
   const [passwordInput, setPasswordInput] = useState('');
@@ -26,17 +29,28 @@ export const usePhototalkAction = ({
 
   const { mutate: deleteByAdmin } = useDeletePhototalkByAdmin();
   const { mutate: deleteByGuest } = useDeletePhototalkByGuest();
+  // const { mutate: deleteS3Image } = useDeletePhototalkS3Image();
 
   const openModal = (photoTalk: PhotoTalk, action: ActionMode) => {
     setSelectedPhotoTalk(photoTalk);
     setActionMode(action);
     setModalOpen(true);
   };
+
   const closeModal = () => {
     setModalOpen(false);
     setPasswordInput('');
     setSelectedPhotoTalk(null);
     setActionMode(null);
+  };
+
+  const checkPassword = (phototalk: PhotoTalk): boolean => {
+    if (phototalk.password !== passwordInput) {
+      alert('비밀번호가 일치하지 않습니다.');
+      return false;
+    }
+
+    return true;
   };
 
   const confirmAction = () => {
@@ -46,7 +60,9 @@ export const usePhototalkAction = ({
       if (actionMode === ACTION_MODE.DELETE) {
         deleteByAdmin(selectedPhotoTalk.id!, {
           onSuccess: () => {
+            // deleteS3Image(selectedPhotoTalk.id!);
             onDelete?.(selectedPhotoTalk);
+            refetch?.();
             closeModal();
           },
         });
@@ -55,20 +71,30 @@ export const usePhototalkAction = ({
 
     if (mode === USER_MODE.GUEST) {
       if (actionMode === ACTION_MODE.EDIT) {
+        if (!checkPassword(selectedPhotoTalk)) {
+          return;
+        }
+
         onEdit?.(selectedPhotoTalk);
         closeModal();
       }
 
       if (actionMode === ACTION_MODE.DELETE) {
+        if (!checkPassword(selectedPhotoTalk)) {
+          return;
+        }
+
         deleteByGuest(
           {
             id: selectedPhotoTalk.id!,
             name: selectedPhotoTalk.name,
-            password: selectedPhotoTalk.password,
+            password: passwordInput,
           },
           {
             onSuccess: () => {
+              // deleteS3Image(selectedPhotoTalk.id!);
               onDelete?.(selectedPhotoTalk);
+              refetch?.();
               closeModal();
             },
           },
